Changed how video is getting downloaded
